requirements.txt
Scrapy
ipython
python-dateutil
